Don't migrate tables from other engines if they're not in the DB (e.g., in test)

+++ db/migrate/20160916234520_update_to_utf88mb4_stash_engine4.rb
@@ -25,6 +25,9 @@ class UpdateToUtf88mb4StashEngine4 < ActiveRecord::Migration
   private
 
   def set_utf8mb4(table_name, col_name, collation = 'utf8mb4_unicode_ci')
+    warn "Table '#{table_name}' does not exist" unless table_exists?(table_name)
+    return unless table_exists?(table_name)
+
     # index_exists? only seems to work if the index was created in rails migrations and index is named a certain way
     has_index = index_exists?(table_name.intern, col_name.intern)
     remove_index(table_name.intern, column: col_name.intern ) if has_index
